Refactor TrackingValidation helpers and update documentation
